requirements.txt
fastapi
uvicorn[standard]
peewee
mysql-connector-python
bcrypt
face-recognition
opencv-python
ultralytics
numpy